refactor auth0 login
