+++ controllers/authUser/authControllers.js
@@ -1,9 +1,9 @@
 const bcrypt = require('bcrypt');
-const db = require('../../models/userModel');
+// const db = require('../../models/userModel');
 const { generateToken } = require('../../utils/generateToken');
 const requestHandler = require('../../utils/requestHandler');
 const Mailer = require('../../utils/mailHandler');
-const server = require('../../api/server');
+// const server = require('../../api/server');
 const checkItem = require('../../utils/checkInputs');
 const userModel = require('../../models/userModel');
 
@@ -39,29 +39,29 @@ const Login = (req, res) => {
   }
 };
 
-const getAuthToken = async (req, res) => {
-  try {
-    const data = server.locals;
-    if (!data) {
-      res.status(400).json({
-        statusCode: 400,
-        message: 'Authentication Failed'
-      });
-    }
-    const user = await db.createOrFindUser(data);
-    if (user) {
-      req.user = server.locals;
-      generateToken(
-        res,
-        200,
-        `${req.user.authType} Login was successfull`,
-        user
-      );
-    }
-  } catch (error) {
-    return error;
-  }
-};
+// const getAuthToken = async (req, res) => {
+//   try {
+//     const data = server.locals;
+//     if (!data) {
+//       res.status(400).json({
+//         statusCode: 400,
+//         message: 'Authentication Failed'
+//       });
+//     }
+//     const user = await db.createOrFindUser(data);
+//     if (user) {
+//       req.user = server.locals;
+//       generateToken(
+//         res,
+//         200,
+//         `${req.user.authType} Login was successfull`,
+//         user
+//       );
+//     }
+//   } catch (error) {
+//     return error;
+//   }
+// };
 
 const passwordReset = async (req, res) => {
   try {
@@ -128,7 +128,7 @@ const confirmEmail = async (req, res) => {
 module.exports = {
   register,
   Login,
-  getAuthToken,
+  // getAuthToken,
   passwordReset,
   newPassword,
   confirmEmail
